Enhancement: Combine traits into TestTrait
