+++ README.md
@@ -67,17 +67,63 @@ class BazTest extends \PHPUnit_Framework_TestCase
 }
 ```
 
-If you want to yield values from multiple data providers, use the `DataProviderTrait`:
+### Helpers
+
+If you want to make use of a few test helpers, use the `Refinery29\Test\Util\TestHelper` trait!
+
+#### Providing data from an array of values
+
+Use `provideData()` to provide values from an array of values:
+
+```php
+namespace Foo\Bar\Test;
+
+use Foo\Bar\Baz;
+use Refinery29\Test\Util\TestHelper;
+
+class BazTest extends \PHPUnit_Framework_TestCase
+{
+    use TestHelper;
+    
+    /**
+     * @dataProvider providerInvalidValue
+     * 
+     * @param mixed $name
+     */
+    public function testConstructorRejectsInvalidValue($name)
+    {
+        $this->expectException(\InvalidArgumentException::class);
+        
+        new Baz($name);
+    }
+    
+    /**
+     * @return \Generator
+     */
+    public function providerInvalidValue()
+    {
+        return $this->provideData([
+            'foo',
+            'bar',
+        ]);
+    }
+}
+```
+
+#### Providing data from multiple concrete data providers
+
+Use `provideDataFrom()` to provide values from multiple concrete data providers:
 
 ```php
 namespace Foo\Bar\Test;
 
 use Foo\Bar\Baz;
 use Refinery29\Test\Util\DataProvider;
+use Refinery29\Test\Util\TestHelper;
 
 class BazTest extends \PHPUnit_Framework_TestCase
 {
-    use DataProvider\DataProviderTrait;
+    use TestHelper;
     
     /**
      * @dataProvider providerInvalidValue
@@ -104,18 +150,21 @@ class BazTest extends \PHPUnit_Framework_TestCase
 }
 ```
 
-If you want to yield values from multiple data providers, but want to combine 
-with some other values, use the `DataProviderTrait` along with the `Elements` data provider:
- 
+#### Providing data from concrete data providers and arrays of data
+
+Use `provideDataFrom()` along with the `Elements` data provider to 
+combine concrete data providers with data from an array of values:
+
 ```php
 namespace Foo\Bar\Test;
 
 use Foo\Bar\Baz;
 use Refinery29\Test\Util\DataProvider;
+use Refinery29\Test\Util\TestHelper;
 
 class BazTest extends \PHPUnit_Framework_TestCase
 {
-    use DataProvider\DataProviderTrait;
+    use TestHelper;
     
     /**
      * @dataProvider providerInvalidValue
@@ -136,33 +185,66 @@ class BazTest extends \PHPUnit_Framework_TestCase
     {
         return $this->provideDataFrom(
             new DataProvider\InvalidString(),
+            new DataProvider\BlankString(),
             new DataProvider\Elements([
-                'foo',    
-                'bar',    
+                'foo',
+                'bar',
             ])
         );
     }
 }
 ```
 
-### Faker GeneratorTrait
+### Faker
 
-If you need fake data in your tests, the `GeneratorTrait` comes in very handy, as it
-yields an instance of `Faker\Generator` (see [`fzaninotto/faker`](https://github.com/fzaninotto/Faker)):
+If you need fake data in your tests, lazily fetch an instance of `Faker\Generator` 
+(see [`fzaninotto/faker`](https://github.com/fzaninotto/Faker)) using `getFaker()`
 
 ```php
 namespace Foo\Bar\Test;
 
 use Foo\Bar\Baz;
-use Refinery29\Test\Util\Faker\GeneratorTrait;
+use Refinery29\Test\Util\TestHelper;
 
 class BazTest extends \PHPUnit_Framework_TestCase
 {
-    use GeneratorTrait;
+    use TestHelper;
 
     public function testConstructorSetsName()
     {
         $faker = $this->getFaker();
+
+        $name = $faker->name;
+        $date = $faker->dateTime;
+
+        $baz = new Baz(
+            $name,
+            $date
+        );
+
+        $this->assertSame($name, $baz->name());
+        $this->assertNotSame($date, $baz->date());
+        $this->assertEquals($date, $baz->date());
+    }
+}
+```
+
+If you need an instance of `Faker\Generator` with a locale other than 
+the default (`en_US`), pass in the locale to `getFaker()`:
+
+```php
+namespace Foo\Bar\Test;
+
+use Foo\Bar\Baz;
+use Refinery29\Test\Util\TestHelper;
+
+class BazTest extends \PHPUnit_Framework_TestCase
+{
+    use TestHelper;
+
+    public function testConstructorSetsName()
+    {
+        $faker = $this->getFaker('de_DE');
 
         $name = $faker->name;
         $date = $faker->dateTime;
