+++ src/TestHelper.php
@@ -1,0 +1,19 @@
+<?php
+
+/*
+ * Copyright (c) 2016 Refinery29, Inc.
+ *
+ * For the full copyright and license information, please view
+ * the LICENSE file that was distributed with this source code.
+ */
+
+namespace Refinery29\Test\Util;
+
+use Refinery29\Test\Util\DataProvider\DataProviderTrait;
+use Refinery29\Test\Util\Faker\GeneratorTrait;
+
+trait TestHelper
+{
+    use DataProviderTrait;
+    use GeneratorTrait;
+}
