+++ test/DataProviderFake.php
@@ -7,7 +7,7 @@
  * the LICENSE file that was distributed with this source code.
  */
 
-namespace Refinery29\Test\Util\Test\DataProvider;
+namespace Refinery29\Test\Util\Test;
 
 use Refinery29\Test\Util\DataProvider\DataProviderInterface;
