+++ test/TestTraitTest.php
@@ -7,22 +7,15 @@
  * the LICENSE file that was distributed with this source code.
  */
 
-namespace Refinery29\Test\Util\Test\Faker;
+namespace Refinery29\Test\Util\Test;
 
 use Faker\Generator;
-use Refinery29\Test\Util\Faker\GeneratorTrait;
 use Refinery29\Test\Util\Faker\Provider;
+use Refinery29\Test\Util\TestHelper;
 
-class GeneratorTraitTest extends \PHPUnit_Framework_TestCase
+class TestTraitTest extends \PHPUnit_Framework_TestCase
 {
-    use GeneratorTrait;
-
-    public function testCanGetFaker()
-    {
-        $faker = $this->getFaker();
-
-        $this->assertInstanceOf(Generator::class, $faker);
-    }
+    use TestHelper;
 
     public function testCanGetFakerStatically()
     {
@@ -129,5 +122,107 @@ class GeneratorTraitTest extends \PHPUnit_Framework_TestCase
         }
 
         return $data;
+    }
+
+    public function testProvideDataYieldsValues()
+    {
+        $faker = $this->getFaker();
+
+        $values = array_combine(
+            $faker->unique()->words(5),
+            $faker->unique()->words(5)
+        );
+
+        $generator = $this->provideData($values);
+
+        $this->assertGeneratorYieldsValues($values, $generator);
+    }
+
+    public function testProvideDataFromYieldsValuesFromDataProviderWithArrayOfValues()
+    {
+        $faker = $this->getFaker();
+
+        $values = array_combine(
+            $faker->unique()->words(5),
+            $faker->unique()->words(5)
+        );
+
+        $dataProvider = new DataProviderFake($values);
+
+        $generator = $this->provideDataFrom($dataProvider);
+
+        $this->assertGeneratorYieldsValues($values, $generator);
+    }
+
+    public function testProvideDataFromYieldsValuesFromDataProviderWithTraversableYieldingValues()
+    {
+        $faker = $this->getFaker();
+
+        $values = array_combine(
+            $faker->unique()->words(5),
+            $faker->unique()->words(5)
+        );
+
+        $dataProvider = new DataProviderFake($this->traversableFrom($values));
+
+        $generator = $this->provideDataFrom($dataProvider);
+
+        $this->assertGeneratorYieldsValues($values, $generator);
+    }
+
+    public function testProvideDataFromYieldsValuesFromMultipleDataProviders()
+    {
+        $faker = $this->getFaker();
+
+        $valuesOne = array_combine(
+            $faker->unique()->words(5),
+            $faker->unique()->words(5)
+        );
+
+        $valuesTwo = array_combine(
+            $faker->unique()->words(5),
+            $faker->unique()->words(5)
+        );
+
+        $generator = $this->provideDataFrom(
+            new DataProviderFake($this->traversableFrom($valuesOne)),
+            new DataProviderFake($this->traversableFrom($valuesTwo))
+        );
+
+        $values = array_merge(
+            $valuesOne,
+            $valuesTwo
+        );
+
+        $this->assertGeneratorYieldsValues($values, $generator);
+    }
+
+    /**
+     * @param $generator
+     * @param $values
+     */
+    private function assertGeneratorYieldsValues($values, $generator)
+    {
+        $this->assertInstanceOf(\Traversable::class, $generator);
+
+        $expected = array_map(function ($value, $key) {
+            return [
+                $key => $value,
+            ];
+        }, array_values($values), array_keys($values));
+
+        $this->assertSame($expected, iterator_to_array($generator));
+    }
+
+    /**
+     * @param array $values
+     *
+     * @return \Generator
+     */
+    private function traversableFrom(array $values)
+    {
+        foreach ($values as $key => $value) {
+            yield $key => $value;
+        }
     }
 }
